requirements.txt
python-dotenv
google-generativeai
supabase
psycopg2-binary